Testing the DockerMain function
As part of the refactor of the docker system, i have to run a FUNCTION in parallel so the runtimes respects the Task.Delay function.

Note: Also removed the "from-file" command from the docker-compose file, the docker system/mode already does that by default, it doesn't even read the Configuration variable regarding this feature.

Signed: Roxanne